Updated error reporting
Updated return type of cipher from `string` to `bool, string`
The bool value will be used to see if the cipher succeeded.

Updated chec_error(bool, string, Options) to return a bool instead of exiting.
This bool will be used to see if it errored. If the result is false,
there was no error, if it is true, there was an error. This way the
cipher can print out as many errors it wants before returning with a
faliure. Preferably the cipher will show as many errors as it can
before the error would affect the cipher, etc.
(e.g. you check if the input is invalid, if there are more error
checks before the input is used then you can check all of them before
exiting.)

Updated main function to handle the new return types of the ciphers.
If there is an error then it prints "Cipher had errors. Exiting." this
message may make more sense in the future when we'll implement cipher
chaining and exiting early due to an error will be nececary.

I also updated how the result will be printed, previously the output
was just printing the result, but now the output will be:
"Result | 'cipher output'"
The ' are there to make it easier to see what is whitespace from the
result in the cases where the cipher result can contain whitespace.